refactor: Clean HOC calls
The file `allContacts.js` exports partial applied functions
(`const g = x => f(x)` is the same as `const g = f`).
Destructuring `props` is a bad habit that complicates reading the code:
we do not know which variables come from outside and
which are under the responsibility of our component.
The HOC `ContactAppWithLoading` did not handle loading, so it has been removed.