fix: Configure jest to handle a regression
With jest and jsdom version xxx was introduced the need to set
an url in the jest.config.js file.
cf. https://github.com/jsdom/jsdom/issues/2304

And create-cozy-app do not include jest config in its cozy-scripts package.
So even if the issue were fixed, app relying on it must fix it again.
cf. https://github.com/CPatchane/create-cozy-app/pull/706

But how this was introduced? By a bot that seems not very
capable to catch the error introduced by the dependance of a dependance update.
cf. https://github.com/cozy/cozy-contacts/pull/216