fix: Unfreeze the application when selecting a contact
We decided to use `componentWillReceiveProps` to synchronize unsynchronized data.
See https://github.com/cozy/cozy-contacts/pull/87/files\#diff-b5b123b65f648395a7716fa02882b0b0R103

But with the migration to React 16, this function became deprecated, we decided to use
`componentDidUpdate`. See https://github.com/cozy/cozy-contacts/pull/215/\#discussion_r208131374

In the application, the state (the contacts array) is not under the responsibility of the
application but the library `cozy-client`. If we need to delete an item of this array
we call the library and the props of the application changed and the render is called to
update the UI. But if we want to change a property of an item, the modal did not get the change
because we duplicate the item: one copy is passed as props to the modal, the original stay
in the contacts array. That is a smell and we should not multiply source of truth because
it will lead to desynchronization, sooner or later.
`setState` in `componentDidUpdate` could cause a infinite loop of rendering.

As a quick fix, we commit these changes, but we need to review the whole architecture to
not have to duplicate data to avoid desynchronisation.