fix: Reset input to fire onChange on Chrome
On Chrome, `onChange` is not fired when the selected file
is the same (I guess there is no real change for webkit).
See https://stackoverflow.com/questions/6623310/input-file-onchange-event-not-being-fired-in-chrome
Using a controlled component to handle input state is a good way
in React to reset the input value when needed.

fix 252