fix: :fire: Avoid the glitch while loading the app
The users may have a flash of unstyled content without this fix:
browsers don't wait that the CSSOM is ready to execute <script> with the
defer attribute. It means that they can render before the CSS have been
parser, and SVG icons are often rendered on the full screen in that
case.