feat: Remove useless babel-transformer
This babel-transformer does not seem to be useful as babel-jest will
picks up the babel.config.js by itself.
It actually caused a problem since it tried to load babel-jest that
is supposed to be hidden behind cozy-scripts. Since babel-jest was
not available in the root node_modules, but in
node_modules/cozy-scripts/node_modules, the require would not work