feat: Hot reload without error
We cannot use memoize for setupAppContext in the same file as the
module.hot.accept otherwise the setupAppContext function is redefined
and memoization does not work. When the function is in its own module,
the function is not redefined, and memoization works correctly : the
function is not called again, thus initBar is not called again,
preventing a "Target is not a DOM element" error.