let user pick a leading letter to filter by
This commit adds a component that displays all alphabet's letters. A
user can then click on any of those letter to filter contacts whose name
start with the selected letter. Contacts' last names are checked against
the letter, and if only a first name exists it will be checked instead.
This also adds a button "All" to revert to showing all contacts.