fix: Disable global styles
With MUI V4, it is possible to generate deterministic class names.
In the case of multiple react roots, it is necessary to disable this
feature. Since we have the cozy-bar root, we need to disable the
feature.

https://material-ui.com/styles/api/#stylesprovider

Otherwise generated class names clash with others and some text become
the wrong color during the usage of the app (especially visible on
mobile).