fix: Add realtime to contacts
When we arrive on the app, we can launch
the service to keep the index up to date.
By doing so, the service can upgrate the
contact's documents.

In that case, if we try to edit / delete a contact
that have been edited by the service, we will get
a 409 error since the app doesn't have the right
version of the document.

To fix this issue, let's connect to the realtime
to get the latest revision if a document is changed
by something.