chore: Fix integration of Sentry
Sentry could not intercept error logs because
`cozy-logger` did not do them.

To intercept the error logs, you must add the
`CaptureConsole` integration of Sentry,
this is necessary because otherwise all the error logs
in the `try/catch` are invisible.

And added `ignoreErrors`, to avoid polluting Sentry
with React `warnings` (called with "console.error").