fix: Correct order of contacts
We did a manual sort and put "me" first in contacts list
if the service was never launched.
If the service has already been launched, we retrieve
the contacts such as (because it is the query which does
the asc sorting), but in this case we must also put "me" first.