feat(formValuesToContact): Clean associated data
Clean the associated data so that cozy-client does not
hydrate the `relationships` attribute of the document
with the attributes defined in the schema.

Today a document hydrated via `.include` of a query is
with all the `relationships` defined in the app schema.

Therefore, when cozy-client dehydrates the document it
hydrates the `relationships` attribute of the document
with these (even if empty).