fix: Creation of a contact when selecting a linked contact
When we wanted to create a new contact from
the linked contact addition,
the form component was rendered because the
creation of the new contact triggers useQueries
higher up the React tree.

This re-rendering caused the form to lose any changes
it had already made, such as adding the new linked contact.